+++ endorse/lib/annotations.dart
@@ -3,7 +3,7 @@ export 'src/endorse/case.dart';
 export 'src/endorse/validations.dart';
 export 'src/endorse/validate_value.dart';
 export 'src/endorse/validate_list.dart';
-export 'src/endorse/validate_map.dart';
+export 'src/endorse/validate_class.dart';
 export 'src/endorse/result_object.dart';
 export 'src/endorse/class_result.dart';
 export 'src/endorse/value_result.dart';