+++ endorse/lib/src/endorse/result_object.dart
@@ -1,16 +1,10 @@
-
+import 'package:endorse/src/endorse/validation_error.dart';
 
 abstract class ResultObject {
-  Object get value;
-  Object get errors;
-  bool get isValid;
+  Object get $value;
+  String get $fieldName;
+  bool get $isValid;
+  bool get $isNotValid;
+  List<ValidationError> get $errors;
+  Object get $errorsJson;
 }
-
-
-
-
-
-
-
-
-